# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Clinic booking API (learning)
  description: |
    Hand-maintained OpenAPI spec. Update this file when routes change; UI at `/api/docs`.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local (default PORT from `.env`)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register patient or doctor
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: JWT and public user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Doctor record not found (doctor registration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: |
            `EMAIL_TAKEN` — an **active** account already uses this email.
            `EMAIL_RETIRED` — this email belongs to a **soft-deleted** account (unique email prevents reuse in the demo DB).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT and public user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access + refresh pair
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New JWTs and public user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current session profile
      description: Returns the signed-in user and, for doctors, the linked directory row (name, specialty).
      operationId: getAuthMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User plus optional doctor directory profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    delete:
      tags: [Auth]
      summary: Soft-delete current patient or doctor account (demo)
      operationId: deleteAuthMe
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account closed; session should be cleared on the client
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Role not allowed to close via this endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use Authorization header — `Bearer <token>` from `/api/v1/auth/login`
...
  schemas:
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [patient, doctor]
          default: patient
        doctorRecordId:
          type: integer
          minimum: 1
          description: Required when role is doctor — must match `doctors.id`
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: Short-lived access JWT (`Authorization` header)
        refreshToken:
          type: string
          description: Long-lived refresh JWT for `POST /api/v1/auth/refresh`
        user:
          $ref: "#/components/schemas/PublicUser"
    AuthMeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
        doctorProfile:
          description: Linked clinic directory row when the user is a doctor; omitted or null for patients
          nullable: true
          $ref: "#/components/schemas/Doctor"
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        name:
          type: string
        doctorRecordId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: string
          nullable: true
...
    ErrorBody:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        requestId:
          type: string
          format: uuid